Make raw_distance function commutative  (#47)
* Make raw_distance function commutative

Added proptest testing that raw_distance is commutative, which failed
Modified the raw_distance function to be more concise, and pass the
commutativity property test. Adjust jaccard of two empty sets to be 1,
not 0 (as I believe is standard).

* remove extraneous test case

* update `raw_distance` documentation, preconditon

Rewrite the documentation for the `raw_distance` function, including the
fact that its arguments must be in sorted order. Add a debug-only
precondition which checks that they are sorted.

* compactify tests